Update wp-create-legacy.sh
1. wait 1min at maximum for wp-config.php and database connection to be ready
2. Fix hardcoded OSF name

+++ scripts/wp-create-legacy.sh
@@ -8,6 +8,21 @@ function conflict_fail() {
   exit 1
 }
 
+function wait_for_something() {
+    local _what
+    _what=2
+    _wait_time=0
+    _MAX_WAIT_TIME=60
+    while ! $1 ; do
+        sleep 1
+        _loop=$((_loop + 1))
+        if [ $_loop -gt 60 ]; then
+            echo "$_what doesn't exist after waiting $_MAX_WAIT_TIME seconds, something wrong!"
+            exit 1
+        fi
+    done
+}
+
 # setup environment variables...
 DB_PASSWORD="$(head -c 32 /dev/urandom | base64 | tr -cd "[a-zA-Z0-9]")"
 PREFIX=${1:-blog}
@@ -32,14 +47,13 @@ podman container exists "${PREFIX}-fpm" && conflict_fail container "${PREFIX}-fp
 podman container exists "${PREFIX}-nginx" && conflict_fail container "${PREFIX}-nginx"
 
 # create storage volumes...
-echo ""
-echo "Creating storage volumes"
+printf "Creating storage volumes\n"
 podman volume create "${PREFIX}-var-www-html"
 podman volume create "${PREFIX}-mariadb"
 podman volume create "${PREFIX}-nginx-conf"
 
 # create resources (optionally publish on port 80)...
-echo "Creating pod ${PREFIX} which will listen on port $PORT"
+printf "\nCreating pod ${PREFIX} which will listen on port $PORT\n"
 podman pod create --name "${PREFIX}" --publish $PORT:$PORT/tcp
 # create a secret to store mysql password...
 printf "%s" "${DB_PASSWORD}" | podman secret create "${PREFIX}-mysql-root" -
@@ -65,10 +79,9 @@ echo "server {
 }" > $(podman volume inspect -f '{{ .Mountpoint }}' "${PREFIX}-nginx-conf")/blog.conf
 
 
-echo ""
-echo "Starting mariadb, fpm and nginx instances"
+printf "\nStarting mariadb, fpm and nginx instances\n"
 # spawn mariadb instance...
-podman run --pod ${PREFIX} --rm -d --label io.containers.autoupdate=registry \
+podman run --pod ${PREFIX} -d --label io.containers.autoupdate=registry \
   --name "${PREFIX}-db" \
   --secret "${PREFIX}-mysql-root" \
   --env MARIADB_ROOT_PASSWORD_FILE="/run/secrets/${PREFIX}-mysql-root" \
@@ -76,7 +89,7 @@ podman run --pod ${PREFIX} --rm -d --label io.containers.autoupdate=registry \
   --volume "${PREFIX}-mariadb:/var/lib/mysql" \
   docker.io/library/mariadb:latest
 # spawn fpm instance to execute requests...
-podman run --pod ${PREFIX} --rm -d --label io.containers.autoupdate=registry \
+podman run --pod ${PREFIX} -d --label io.containers.autoupdate=registry \
   --name "${PREFIX}-fpm" \
   --secret "${PREFIX}-mysql-root" \
   --env WORDPRESS_DB_HOST=127.0.0.1 \
@@ -86,20 +99,23 @@ podman run --pod ${PREFIX} --rm -d --label io.containers.autoupdate=registry \
   --volume "${PREFIX}-var-www-html:/var/www/html" \
   docker.io/library/wordpress:fpm-alpine
 # spawn nginx instance to serve requests...
-podman run --pod ${PREFIX} --rm -d --label io.containers.autoupdate=registry \
+podman run --pod ${PREFIX} -d --label io.containers.autoupdate=registry \
   --name "${PREFIX}-nginx" \
   --volume "${PREFIX}-nginx-conf:/etc/nginx/conf.d" \
   --volume "${PREFIX}-var-www-html:/var/www/html" \
   docker.io/library/nginx:alpine
 
 
-# wait for wp-config.php and mariadb to be ready
-sleep 5
+# wait for wp-config.php to be ready
+function wait_for_wp_config() {
+    podman exec -it "${PREFIX}-fpm" sh -c "test -f /var/www/html/wp-config.php"
+}
+wait_for_something wait_for_wp_config "/var/www/html/wp-config.php"
 
 podman exec -it $PREFIX-nginx sed -i "/WORDPRESS_TABLE_PREFIX/a define('WP_HOME', 'http://localhost:$PORT');\ndefine('WP_SITEURL', 'http://localhost:$PORT');" /var/www/html/wp-config.php
 podman exec -it $PREFIX-nginx sed -i 's/^\$table_prefix.*$/\$table_prefix="wpnf_";/' /var/www/html/wp-config.php
 
-echo "Importing OSF's database"
+printf "\nImporting OSF's database\n"
 CONTAINER_DB_PATH=/var/lib/mysql/osfdb.sql
 if grep -E -qs ".zip$" <<< "$3"; then
     tmp_db=$(sed -E -n "s/(.*).zip$/\1/p" <<< "$3")
@@ -111,19 +127,34 @@ fi
 
 tmp_db=/tmp/"$(basename "$tmp_db")"
 
-podman cp "$tmp_db" OSF-db:$CONTAINER_DB_PATH
+if ! podman cp "$tmp_db" ${PREFIX}-db:$CONTAINER_DB_PATH; then
+    echo "Failed to copy database $tmp_db to container"
+    rm -f $tmp_db
+    exit 1
+fi
 rm -f $tmp_db
 
-PASSWORD=`podman exec "${PREFIX}-db" cat "/run/secrets/${PREFIX}-mysql-root"`
 
-podman exec -it "${PREFIX}-db" sh -c "mariadb -uroot -p${PASSWORD} < $CONTAINER_DB_PATH"
+function wait_for_db() {
+    podman exec -it "${PREFIX}-db" sh -c "mariadb -uroot -p${DB_PASSWORD} -e exit &> /dev/null"
+    #podman exec -it "${PREFIX}-db" sh -c "test -e /run/mysqld/mysqld.sock "
+}
+
+# Wait for database server to be ready
+# somehow "wait_for_something wait_for_db" doesn't work as expected and "sleep
+# 5" is still needed
+sleep 5
+wait_for_something wait_for_db "Database connection"
+
+if podman exec -it "${PREFIX}-db" sh -c "mariadb -uroot -p${DB_PASSWORD} < $CONTAINER_DB_PATH"; then
 podman exec -it "${PREFIX}-db" sh -c "rm -f $CONTAINER_DB_PATH"
+fi
 
 if [[ -d "$WP_CONTENT" ]]; then
-    echo "Copy OSF's wp-content which includes theme, plugins to container"
+    printf "\nCopying OSF's wp-content which includes theme, plugins and some uploads to container\n"
     podman cp "$WP_CONTENT" "$PREFIX-nginx":/var/www/html/
-    echo "Fixing permission of copied files"
-    wp_content_dir=$(podman volume inspect OSF-var-www-html| sed -En 's/^\s+"Mountpoint": "(.*)",$/\1/p')/wp-content
+    printf "\nFixing permission of copied files\n"
+    wp_content_dir=$(podman volume inspect -f '{{ .Mountpoint }}' "${PREFIX}-var-www-html")/wp-content
     uid=$(stat -c '%u' "${wp_content_dir}/plugins/akismet")
     if [[ -z $uid ]]; then
         echo "Failed to get uid of pre-isntalled akismet"
